pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Łukasz Gendosz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
uvicorn = "^0.30.5"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
dependency-injector = "^4.41.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
orjson = "^3.10.7"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
